cli: add vault subcommand
The cli can act as an client to the uno api. This begins the journey.
vault subcommand can PUT and GET an arbitrary vault. The logic for
executing http requests lives in the uno lib right now. It might make
more sense in its own crate or even directly in the cli, but it's fine
here for now.