ffi: add script to build xcframework
We need to use a custom target spec for the iOS simulator. More details:

  https://github.com/rust-lang/rust/issues/79408

Since that issue, all the targets except the x86_64 simulator have been
updated (one landed last night). There is a PR that updates the
simulator:

  https://github.com/rust-lang/rust/pull/87699

Once that PR is merged we can remove the `x86_64-apple-ios7.0.json`
custom target spec and use the builtin.