ffi: fix naming and a few types
Switch to not allocating #[repr(C)] structs. The caller allocates memory
as we just write the memory instead of writing a leaked boxed pointer.
The caller is still responsible for freeing the resources associated
with the structs since many of them contain pointers to other memory we
own. I may reverse this in the future to make it more clear/consistent
such that we always return a pointer to memory we allocate when the user
is required to later pass that struct to an uno_free function, but for
now this is simpler and it hasn't proven confusing yet.