ffi: add xcode12 and xcode13 compat builds
sumo.sh builds an UnoRust.xcframework with the latest nightly. The
problem is that xcode uses an older version of llvm and barfs on some
unknown attributes in the binary, which is sad. To fix this we must
build with a nightly version where the default llvm version matches
the version that xcode uses. For Xcode12, that's 1.51, for Xcode13,
that's up to 1.55. Currently the nightlys are on 1.56.

Additionally, aarch64-apple-ios-simulator (M1) is no longer supported
on Xcode12. It does work on Xcode13. This is super annoying because
the issue seems to be a "regression" in `xcodebuild xcframework`. It
now complains when we try to add the aarch64 simulator slice built
using a custom rust target. This is sad because it previously worked.

Xcode13 is only a few weeks out so no need to fuss too much, though.