xcf: add dedicated xcframework crate
We're going to have to build binaries for different platforms beyond
macOS/iOS and they may require different bindgen or need to be built
in entirely separate repos, etc. Isolate the xcodebuild and sumo
xcframework build scripts and target specs to their own crate. It is
easy enough to do multiple cbindgen runs on a reexported ffi crate.

An example would be a future `jni` crate that contains everything
needed to build an appropriate kotlin package that can interface
via kotlin native or the jni, for Android. The `xcf` and `jni`
crates may get pulled out entirely but it really depends on how hard
it is to get all the build tooling hooked up in gh-actions.