vclock: wip it compiles
Store the vault in a document rather than as raw data. The Vault
document contains the vclock alongside the raw data. Store the doc
as json right now even though it's not the most efficient format.
Parse the vclock header manually until we can add support to serde
for rfc8941. Nest the api under `/v2` because this is a breaking
change to the vault endpoint. We will need to namespace data in
the s3 buckets and also migrade existing docs.

TODO: tests, cli, migrate