cli: add `uno init`
Initialize into a config directory in the user's home environment.
Uno generates and persists the identity seed in the config directory
and future invocataions of `uno seed` display the configured value.
This saves the user needing to pass a `--seed` parameter on every
command invocation.

TODO: add support to other subcommands for reading the persistent
      identity.