api: remove old directory entries when phone changes
If a phone number becomes associated with a new directory entry, then
remove the old one in the process. This is more of a housekeeping
concern because all the logic keys off of phone number and not CID.
But this does prevent a client holding on to cid from getting a stale
directory entry.