uno-assistant: add initial groundwork
Call the lambda for now. The end goal is to port the logic from the
DO serverless function into this repo, since we don't have auth on the
serverless endpoint, but there's some entropy in the url and it's not
a huge deal to use it for now while things mature.